refactor: Apply clean architecture
Add domain dir, and infra dir.
