+++ src/adapter.rs
@@ -1,7 +1,6 @@
 use serde_json::Value;
 
-use crate::entity::MinecraftMod;
-use crate::modrinth_api;
+use crate::{domain::entities::minecraft_mod::MinecraftMod, infrastructure::modrinth_api};
 
 pub(crate) async fn lists_projects_versions_to_new_mc_mod_download_url(
     mc_version: &str,
