+++ src/domain/contract/add.rs
