+++ src/domain/contract/latest.rs
