+++ src/domain/entities/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod minecraft_mod;
