+++ src/domain/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod contract;
+pub(crate) mod entities;
