+++ src/infrastructure/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod file_system;
+pub(crate) mod modrinth_api;
