+++ src/service/rm.rs
@@ -1,8 +1,11 @@
-use crate::{contract::rm::RemoveMinecraftMod, service::MinecraftModsService, utils};
+use crate::{
+    domain::contract::rm::RemoveMinecraftMod, infrastructure::file_system,
+    service::MinecraftModsService,
+};
 
 impl RemoveMinecraftMod for MinecraftModsService {
     fn rm(&self, mc_mod_file_name: &str) -> Result<(), Box<dyn std::error::Error>> {
-        utils::remove_mc_mod_by_file_name(mc_mod_file_name)?;
+        file_system::remove_mc_mod_by_file_name(mc_mod_file_name)?;
 
         Ok(())
     }
