+++ src/utils.rs
@@ -1,14 +1,16 @@
 use crate::{
-    entity::{MinecraftMod, MinecraftModVersionUpdate},
-    modrinth_api,
+    domain::entities::minecraft_mod::{MinecraftMod, MinecraftModVersionUpdate},
+    infrastructure::{
+        file_system::{copy_downloaded_mc_mod_to_file_by_file_name, remove_mc_mod_by_file_name},
+        modrinth_api,
+    },
     service::MinecraftModsService,
 };
 use colored::Colorize;
 use futures::future::join_all;
 use rayon::prelude::*;
 use serde_json::Value;
-use std::fs::File;
-use std::{error::Error, fs, io::copy};
+use std::error::Error;
 use walkdir::{DirEntry, WalkDir};
 
 pub(crate) async fn get_latest_version_of_multiple_project(
@@ -67,18 +69,6 @@ pub(crate) async fn update_mc_mod_to_new_version(
     );
 
     Ok(mc_mod.file_name.to_owned())
-}
-
-async fn copy_downloaded_mc_mod_to_file_by_file_name(
-    file_name: &str,
-    mut response: reqwest::Response,
-) -> Result<(), Box<dyn Error + 'static>> {
-    let mut dest = File::create(format!("mods/{file_name}")).unwrap();
-    while let Some(chunk) = response.chunk().await? {
-        copy(&mut chunk.as_ref(), &mut dest).unwrap();
-    }
-
-    Ok(())
 }
 
 async fn download_mc_mod_by_url(
@@ -325,13 +315,6 @@ async fn update_selective(new_mc_mods: &[MinecraftModVersionUpdate]) {
             }
         }
     }
-}
-
-pub(crate) fn remove_mc_mod_by_file_name(
-    mc_mod_file_name: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    fs::remove_file("mods/".to_string() + mc_mod_file_name)?;
-    Ok(())
 }
 
 pub(crate) async fn update_mods_to_support_a_mc_version_ui(
