+++ src/view/latest.rs
@@ -1,4 +1,4 @@
-use crate::{contract::latest::LatestMinecraftMods, service::MinecraftModsService};
+use crate::{domain::contract::latest::LatestMinecraftMods, service::MinecraftModsService};
 
 pub(crate) async fn latest_view(mc_version: &str) {
     let mut mc_mods_service = MinecraftModsService::default();
